Improvements

+++ linux-user-wizard.sh
@@ -1,4 +1,5 @@
 #!/usr/bin/env bash
+#set -x 
 
 if ! [ $(id -u) = 0 ]; then
     echo "Tou need to have root privileges to run this script
@@ -12,7 +13,7 @@ function exiting {
     read exitanswer
     if [ "$exitanswer" = "y" ] || [ "$exitanswer" = "Y" ]
         then
-        exec "./linux-user-wizard.sh"
+         bash ./linux-user-wizard.sh
     elif [ "$exitanswer" = "n" ] || [ "$exitanswer" = "N" ]
         then
     echo "GOOD BYE"
@@ -52,36 +53,60 @@ if [ "$answer" = "1" ]
     then
     echo "Please enter a username"
     read luwuser
-    useradd $luwuser -s /bin/bash
+    
     if [ -d /home/$luwuser ]
         then echo "Homefolder exists do you want to overwrite it? (y/n)"
         read homefolderanswer
         if [ "$homefolderanswer" = "y" ] || [ "$homefolderanswer" = "Y" ]
             then
             rm -rf /home/$luwuser
-            mkdir /home/$luwuser
+            useradd $luwuser -s /bin/bash
             mkdir /home/$luwuser/.ssh
             touch /home/$luwuser/.ssh/authorized_keys
+            ssh-keygen -t rsa
+            mv /root/.ssh/id_rsa* /home/$luwuser
+            cat /home/$luwuser/id_rsa.pub >> /home/$luwuser/.ssh/authorized_keys
+            chown -R $luwuser /home/$luwuser
+            chmod 600 /home/$luwuser/.ssh/authorized_keys
+            exiting
         else
+	    echo "Not creating the user since you want to keep the homefolder"
             echo "Leaving homefolder intact"
-        fi
-    fi
+            echo "Do you still want to create SSH key and put in user's home folder? (y/n)"
+	    read keyans
+	    if [ "$keyans" = "y" ] || [ "$keyans" = "Y" ]
+	        then
+	    	ssh-keygen -t rsa
+            	mv /root/.ssh/id_rsa* /home/$luwuser
+            	cat /home/$luwuser/id_rsa.pub >> /home/$luwuser/.ssh/authorized_keys
+            	chown -R $luwuser /home/$luwuser
+            	chmod 600 /home/$luwuser/.ssh/authorized_keys
+	    fi
+	   
+	   	exiting
+         fi
+   fi
+fi
 
     if [ ! -d /home/$luwuser ]
         then
-            mkdir /home/$luwuser
+	    useradd $luwuser -s /bin/bash 
             mkdir /home/$luwuser/.ssh
-    fi
+	    touch /home/$luwuser/.ssh/authorized_keys
+            ssh-keygen -t rsa
+            mv /root/.ssh/id_rsa* /home/$luwuser
+            cat /home/$luwuser/id_rsa.pub >> /home/$luwuser/.ssh/authorized_keys
+            chown -R $luwuser /home/$luwuser
+            chmod 600 /home/$luwuser/.ssh/authorized_keys
+            exiting    
 
-    touch /home/$luwuser/.ssh/authorized_keys
-    ssh-keygen -t rsa
-    mv /root/.ssh/id_rsa* /home/$luwuser
-    cat /home/$luwuser/id_rsa.pub >> /home/$luwuser/.ssh/authorized_keys
-    chown -R $luwuser /home/$luwuser
-    chmod 600 /home/$luwuser/.ssh/authorized_keys
-    echo "User has been added"
-    exiting
 fi
+
+#if [ ! -f /home/$luwuser/.ssh/authorized_keys ]
+#   then
+#       touch /home/$luwuser/.ssh/authorized_keys
+#fi
+
 
 
 if [ "$answer" = "2" ]
@@ -111,7 +136,7 @@ fi
 if [ "$answer" = "5" ]
     then
     cat /etc/passwd | grep /bin/bash | less
-    exiting
+    ./linux-user-wizard.sh
 fi
 
 
@@ -139,5 +164,7 @@ fi
 
 if [ "$answer" != "1" ] && [ "$answer" != "2" ] && [ "$answer" != "3" ] && [ "$answer" != "4" ] && [ "$answer" != "5" ] && [ "$answer" != "6" ]
     then
-        exec ./linux-user-wizard.sh
+        bash ./linux-user-wizard.sh
 fi
+
+
